Dev Containers
In the process of implementing Dev Containers to make development more streamline.

This would allow you to edit the code base from inside a container not requiring you to download dependencies locally on your machine.

Currently I have just created the .devcontainer folder and devcontainer.json.

Edited devcontainer.json file to create container you would work in. It is currently not working. I believe the issue is due to the workspaceFolder having the wrong path. I might need to adjust the docker compose file in the root directory.