Worked on Frontend Functionality and Design
Finished:
I move repeated functions to a function called utils.tsx.

I changed how age is calculated. Now the program access for the birthdate of the patient and calculates it.

Fixed some styling on some buttons that were invisible. Now they are visible as a darkish blue color and when you hover over them they get slightly darker.

Finished the edit button. Changes information and displays information on the user. Nothing is saved though. Still waiting on the backend sql lite server to save the information.

Users are now able to select tests for patients and go to the VideoRecording.tsx page.

To Do:
Change quick add button to upload patient data button as a mini pop which uses a dropdown list to select patients.

research how d3 library.